TD-378-3: Fix some UI
