+++ src/components/common/input-autosize/style.module.scss
@@ -10,5 +10,5 @@
 }
 
 .title {
-  @apply align-middle text-xl font-semibold;
+  @apply align-middle text-h3 font-semibold;
 }
